🐝✅ - day 2 part 1 - solved 🔧 - business logic now supports long instead of int for ids

+++ src/advent.of.code-2025/RiddleSolver.cs
@@ -18,7 +18,7 @@ public class RiddleSolver(IEnumerable<string> riddleInputs)
         return dial.Password;
     }
 
-    public int SolveDay2()
+    public long SolveDay2()
     {
         if (RiddleInputs.Count() > 1)
         {
